Aggiunta di audit log, sicurezza JWT/Keycloak, gestione documenti passeggeri e metodi di pagamento

+++ backend/booking-service/src/main/java/com/tripify/booking_service/dto/BookingResponseDTO.java
@@ -3,13 +3,15 @@ package com.tripify.booking_service.dto;
 import com.tripify.booking_service.entity.BookingStatus;
 import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @Builder
 public class BookingResponseDTO {
     private Long id;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private LocalDateTime bookingDate;
     private BookingStatus status;
